refactor(api): split ingest.js controller into per-responsibility modules (#599)
ingest.js had grown to 609 lines (audit finding C1). Split it by
responsibility into focused routers under routes/ingest/:

  handlers.js    — the generic per-entity ingest endpoints (createIngestHandler)
                   + principals-presence, sync-log, classify-business-role-assignments
  matrixViews.js — /ingest/refresh-views + /ingest/matrix-default-filter
                   + the refreshKeyword / refreshMatrixViews helpers
  helpers.js     — the ingest engine helpers createIngestHandler composes

ingest.js is now a barrel that composes the two routers into the default
export and re-exports the helpers that ingest.*.test.js and bootstrap.js
(refreshMatrixViews) import from ./ingest.js — so those consumers, app.js
and every public path are unchanged. Pure code move.

ingest.js is an OpenAPI-documented router, and the drift guard enumerates
its routes by introspecting the router's layer stack. A barrel that
composes sub-routers via router.use() nests the routes one level down, so
the guard's registeredOps() now recurses into nested sub-routers — a
general fix that enumerates any barrel-composed documented router (the
undocumented splits and crawlers, which re-exports its two routers
directly, are unaffected).

ingest.js had no complexity-baseline entries. Part2's changed-line diff
coverage is 83%. Full API suite green (1786 tests); complexity + jscpd
+ the (updated) OpenAPI-drift guard pass.
